# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List published classes
      description: Returns published classes only, with optional search/filter/sort
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            $ref: "#/components/schemas/ClassLevel"
        - name: category
          in: query
          schema:
            type: string
        - name: instructorId
          in: query
          schema:
            type: string
            format: uuid
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price_asc, price_desc, popular]
        - name: includeAll
          in: query
          description: When true, returns classes of any status (for admin use). Defaults to published only.
          schema:
            type: boolean
      responses:
        "200":
          description: List of published classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassSummary"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassRequest"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassSummary"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /classes/{id}:
    get:
      operationId: getClassById
      tags: [classes]
      summary: Get a published class by ID
      description: Returns a single published class with its modules, or 404 if not found or not published
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Class detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassDetail"
        "404":
          description: Class not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateClass
      tags: [classes]
      summary: Update a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassRequest"
      responses:
        "200":
          description: Class updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassSummary"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Class not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Removed
        "404":
          description: Class not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /instructors:
    get:
      operationId: listInstructors
      tags: [instructors]
      summary: List instructors
      description: Returns all instructors with their published class count
      responses:
        "200":
          description: List of instructors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InstructorSummary"
    post:
      operationId: createInstructor
      tags: [instructors]
      summary: Create an instructor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInstructorRequest"
      responses:
        "201":
          description: Instructor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstructorSummary"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /instructors/{id}:
    put:
      operationId: updateInstructor
      tags: [instructors]
      summary: Update an instructor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInstructorRequest"
      responses:
        "200":
          description: Instructor updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstructorSummary"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Instructor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteInstructor
      tags: [instructors]
      summary: Delete an instructor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Removed
        "404":
          description: Instructor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /classes/recommended:
    get:
      operationId: listRecommendedClasses
      tags: [classes]
      summary: List recommended classes for a user
      description: Returns published classes excluding ones already in the user's cart or enrollments
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 3
      responses:
        "200":
          description: Recommended classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassSummary"
...
components:
  schemas:
...
    ClassLevel:
      type: string
      enum: [pemula, menengah, lanjutan]
    InstructorSummary:
      type: object
      required: [id, name, bio, photoUrl, classCount]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        bio:
          type: string
        photoUrl:
          type: string
        classCount:
          type: integer
    ClassInstructor:
      type: object
      required: [id, name, photoUrl]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        photoUrl:
          type: string
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ClassModule:
      type: object
      required: [id, title, orderIndex, durationMinutes]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        orderIndex:
          type: integer
        durationMinutes:
          type: ["integer", "null"]
    ClassDetail:
      type: object
      required:
        - id
        - title
        - description
        - coverImage
        - basePrice
        - discountPrice
        - status
        - level
        - category
        - instructor
        - modules
        - moduleCount
        - totalDurationMinutes
        - youtubePlaylistId
        - gdriveMateriUrl
        - waGroupUrl
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        coverImage:
          type: string
        basePrice:
          type: integer
        discountPrice:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, published]
        level:
          type: ["string", "null"]
          enum: [pemula, menengah, lanjutan, null]
        category:
          type: ["string", "null"]
        youtubePlaylistId:
          type: ["string", "null"]
        gdriveMateriUrl:
          type: ["string", "null"]
        waGroupUrl:
          type: ["string", "null"]
        instructor:
          $ref: "#/components/schemas/InstructorSummary"
        modules:
          type: array
          items:
            $ref: "#/components/schemas/ClassModule"
        moduleCount:
          type: integer
        totalDurationMinutes:
          type: ["integer", "null"]
    ClassSummary:
      type: object
      required:
        - id
        - title
        - description
        - coverImage
        - basePrice
        - discountPrice
        - status
        - level
        - category
        - instructor
        - moduleCount
        - totalDurationMinutes
        - youtubePlaylistId
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        coverImage:
          type: string
        basePrice:
          type: integer
        discountPrice:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, published]
        level:
          type: ["string", "null"]
          enum: [pemula, menengah, lanjutan, null]
        category:
          type: ["string", "null"]
        youtubePlaylistId:
          type: ["string", "null"]
        instructor:
          $ref: "#/components/schemas/ClassInstructor"
        moduleCount:
          type: integer
        totalDurationMinutes:
          type: ["integer", "null"]
    CreateClassRequest:
      type: object
      required: [title, instructorId]
      properties:
        title:
          type: string
        description:
          type: string
        coverImage:
          type: string
        basePrice:
          type: integer
        discountPrice:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, published]
        level:
          type: ["string", "null"]
          enum: [pemula, menengah, lanjutan, null]
        category:
          type: ["string", "null"]
        instructorId:
          type: string
          format: uuid
    UpdateClassRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        coverImage:
          type: string
        basePrice:
          type: integer
        discountPrice:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, published]
        level:
          type: ["string", "null"]
          enum: [pemula, menengah, lanjutan, null]
        category:
          type: ["string", "null"]
        instructorId:
          type: string
          format: uuid
    CreateInstructorRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        bio:
          type: string
        photoUrl:
          type: string
    UpdateInstructorRequest:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
        photoUrl:
          type: string
...